Add support for OpenBSD
On *BSD, there is no group called 'root', therefore just use
the GID of 0 to define the group membership of files and directories
in init.pp and script.pp.

OpenBSD just has:

[ -f /etc/rc.local ] && sh /etc/rc.local

therefore, the shell shebang is not needed here, so update the
template for that case, omitting it when running on OpenBSD.
Further, OpenBSD ls doesn't know about the -v parameter,
therefore make the parameters to ls conditional.
OpenBSD ls doesn't support -v parameter.